requirements.txt
aioredis==1.2.0
amqp==2.5.2
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
billiard==3.6.1.0
boto3==1.9.169
botocore==1.12.169
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
daphne==2.2.5
decorator==4.3.2
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-extensions==1.7.6
django-prometheus==1.1.0
django-redis==4.10.0
django-rest-auth==0.9.5
django-storages==1.7.1
django-timezone-field==3.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docutils==0.14
furl==2.0.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
ipdb==0.11
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.15
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
kombu==4.6.6
lxml==4.3.3
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.1
oauthlib==3.0.1
orderedmultidict==1.0
parsel==1.5.1
parso==0.4.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-crontab==2.4.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
queuelib==1.5.0
redis==3.2.0
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
Scrapy==1.6.0
service-identity==18.1.0
simplegeneric==0.8.1
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.25.3
vine==1.3.0
w3lib==1.20.0
wcwidth==0.1.7
zipp==0.6.0
zope.interface==4.6.0
